Update GameManager with stack-based game rules
The GameManager script now allows game rules to be managed on a stack basis instead of a boolean approach. This modification introduces functions to add and remove game rules from the stack, with the state of the game changing with the number of active rules. The changes also provide the ability to force game rules in Unity Editor.